[DPR-391] Different chart types
