+++ app/views/layout.html
@@ -12,22 +12,17 @@
           items: [
             {
               text: 'Day of the Week',
-              href: '/reports/external-movements-by-weekday',
+              href: '/reports/external-movements-by-weekday-line',
               description: 'External movements by Day of the Week.'
             },
             {
               text: 'Direction',
-              href: '/reports/external-movements-by-direction',
+              href: '/reports/external-movements-by-direction-bar',
               description: 'External movements by Direction.'
             },
             {
-              text: 'Type',
-              href: '/reports/external-movements-by-type',
-              description: 'External movements by Type.'
-            },
-            {
               text: 'Reason',
-              href: '/reports/external-movements-by-reason',
+              href: '/reports/external-movements-by-reason-bar-y',
               description: 'External movements by Reason.'
             }
           ] | setCurrentToActive(currentUrl)
