+++ app/views/reports-chart.html
@@ -28,7 +28,7 @@
 {% endif %}
 
 <h2 class="govuk-heading-m">
-  External movements by {{ head }}
+  {{ title }} by {{ head }}
 </h2>
 <div class="moj-filter-layout">
   <div class="moj-filter-layout__filter sidebar-filter" style="display: none">
@@ -114,7 +114,7 @@
     const ctx = document.getElementById('dataTableChart');
 
     new Chart(ctx, {
-      type: 'bar',
+      type: '{{ chart.type }}',
       data: {
         labels: [
           {% for label in  chart.labels %}
@@ -124,7 +124,7 @@
         datasets: [{
           label: 'Movements',
           data: [{{ chart.data }}],
-          borderWidth: 1
+          borderWidth: {{ chart.borderWidth }}
         }]
       },
       options: {
@@ -134,7 +134,8 @@
           }
         },
         borderColor: '#5694ca',
-        backgroundColor: '#5694ca'
+        backgroundColor: '#5694ca',
+        indexAxis: '{{ chart.indexAxis }}'
       }
     });
   });
